Moved submission module, added code for reloading models
