+++ brainscore/submission/ml_pool.py
@@ -1,9 +1,7 @@
 import warnings
 
-import itertools
-
 from brainscore.utils import LazyLoad
-from submission.utils import UniqueKeyDict
+from brainscore.submission.utils import UniqueKeyDict
 
 
 class ModelLayers(UniqueKeyDict):
@@ -22,26 +20,13 @@ class ModelLayers(UniqueKeyDict):
     def __contains__(self, item):
         return super(ModelLayers, self).__contains__(self._item(item))
 
-
-class ModelLayersPool(UniqueKeyDict):
-    def __init__(self, base_model_pool, model_layers):
-        super(ModelLayersPool, self).__init__()
-        for basemodel_identifier, activations_model in base_model_pool.items():
-            if basemodel_identifier not in model_layers:
-                warnings.warn(f"{basemodel_identifier} not found in model_layers")
-                continue
-            layers = model_layers[basemodel_identifier]
-
-            self[basemodel_identifier] = {'model': activations_model, 'layers': layers}
-
-
 regions = ['V1', 'V2', 'V4', 'IT']
 
 
 class MLBrainPool(UniqueKeyDict):
     def __init__(self, base_model_pool, model_layers):
         super(MLBrainPool, self).__init__()
-
+        self.reload = True
         for basemodel_identifier, activations_model in base_model_pool.items():
             if basemodel_identifier not in model_layers:
                 warnings.warn(f"{basemodel_identifier} not found in model_layers")
@@ -50,7 +35,9 @@ class MLBrainPool(UniqueKeyDict):
 
             from model_tools.brain_transformation import ModelCommitment
             # enforce early parameter binding: https://stackoverflow.com/a/3431699/2225200
+
             def load(identifier=basemodel_identifier, activations_model=activations_model, layers=layers):
+                activations_model.unregister_hooks()
                 brain_model = ModelCommitment(identifier=identifier, activations_model=activations_model,
                                               layers=layers)
                 for region in regions:
